start/requirements.txt
certifi==2019.3.9
Django==2.2
gunicorn==19.9.0
python-decouple==3.1
pytz==2019.1
sqlparse==0.3.0
